Rework Bulk Export Partitioning Support (#7451)
* Work on reindex

* Work

* More work

* Additional tests

* Add changelog

* Resolve FIXMEs

* Docs improvements

* Build fixes

* Test fix

* Test fixes

* Work on tests

* Work on tests

* Test fixes

* Test fixes

* Test fix

* Test fixes

* Version bump

* Test fix

* Add skip indexing

* Work on bulk export

* Work on tests

* Work on bulk export

* About to store request details

* Fix tests

* Test fixes

* Test fix

* Test fixes

* Test fixes

* Spotless

* Merge master

* Test fixes

* FIx tests

* Cleanup

* Test fixes

* Test fixes

* Test fixes

* Cleanup

* Cleanup

* Clanup

* HAPI bump and address review comments

* Cleanup

* Test fix

* Cleanup

* Test fix

+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/dao/BaseHapiFhirDao.java
@@ -982,15 +982,19 @@ public abstract class BaseHapiFhirDao<T extends IBaseResource> extends BaseStora
 				}
 
 				// Extract search params for resource
-				mySearchParamWithInlineReferencesExtractor.populateFromResource(
-						requestPartitionId,
-						newParams,
-						theTransactionDetails,
-						entity,
-						theResource,
-						existingParams,
-						theRequest,
-						thePerformIndexing);
+				boolean skipIndexing =
+						Boolean.TRUE.equals(theResource.getUserData(JpaConstants.RESOURCE_SKIP_INDEXING));
+				if (!skipIndexing) {
+					mySearchParamWithInlineReferencesExtractor.populateFromResource(
+							requestPartitionId,
+							newParams,
+							theTransactionDetails,
+							entity,
+							theResource,
+							existingParams,
+							theRequest,
+							thePerformIndexing);
+				}
 
 				if (CollectionUtils.isNotEmpty(newParams.myLinks)) {
 					setTargetResourceTypeIdForResourceLinks(newParams.myLinks);
@@ -1107,7 +1111,9 @@ public abstract class BaseHapiFhirDao<T extends IBaseResource> extends BaseStora
 		/*
 		 * Indexing
 		 */
-		if (thePerformIndexing) {
+		boolean skipIndexing = theResource != null
+				&& Boolean.TRUE.equals(theResource.getUserData(JpaConstants.RESOURCE_SKIP_INDEXING));
+		if (thePerformIndexing && !skipIndexing) {
 			if (newParams == null) {
 				myExpungeService.deleteAllSearchParams(entity.getPersistentId());
 				entity.clearAllParamsPopulated();
